pyproject.toml
[tool.poetry]
name = "mitmscrape"
version = "0.1.0"
description = ""
authors = ["Max Fan <theinnovativeinventor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mitmproxy = "^5.3.0"
mitmdump = "^1.1.2"
selenium = "^3.141.0"
typer-cli = "^0.0.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
